[Core] Move Context management functions to `Manager`
#428 moved `freezeManagerState` and `thawManagerState` to `Session`
following the removal of the transactions API. This was a stop-gap as
`Session::createContext` already existed so it kind of made sense.

Looking at the bigger picture, `Session` probably isn't the right place
for them:

Context creation/freeze/thaw only actually involves the manager,
the session is just coordinating. A context is tied to the manager
whose state it holds, and so can't be shared/reused with multiple
managers. This means that the current manager of the session may well
change in-between calls, encouraging invalid context use.

Moving context management to Manager eliminates any ambiguity about
the fact that any given context can only be used with the manager that
created the state it holds.

Closes #430
