+++ doc/src/ManagerState.dox
@@ -40,15 +40,13 @@
  *
  * @startuml
  * actor host
- * host -> Session : createContext()
- * Session -> Manager : _createState()
+ * host -> Manager : createContext()
  * Manager -> ManagerInterface : createState(hostSession)
  * ManagerInterface -> ManagerInterface : __createSnapshotToken()
  * ManagerInterface --> Manager : token
- * Manager --> Session : token
- * Session -> Session : new Context()
- * Session -> Session : context.managerInterfaceState = token
- * Session --> host : context
+ * Manager -> Manager : new Context()
+ * Manager -> Manager : context.managerInterfaceState = token
+ * Manager --> host : context
  * ...
  * host -> Manager : register(..., context)
  * Manager -> ManagerInterface : register(..., context, hostSession)
@@ -66,8 +64,8 @@
  * Distributed processing requires the same context to be shared by
  * multiple discrete working processes.
  *
- * The @ref openassetio.hostAPI.Session.Session.freezeContext
- * "Session.freezeContext" and @ref openassetio.hostAPI.Session.Session.thawContext
+ * The @ref openassetio.hostAPI.Manager.Manager.freezeContext
+ * "Manager.freezeContext" and @ref openassetio.hostAPI.Manager.Manager.thawContext
  * "thawContext" methods provide a mechanism by which a @ref host
  * can obtain a portable version of the manager's state token for a
  * given context for distribution to other processes that are logically
@@ -83,18 +81,16 @@
  * == preflight task ==
  * scheduler -> worker : someApp a_job.script ~--preflight --stateFile openassetio.json
  * activate worker
- * activate Session
  * activate Manager
- * worker -> Session : createContext()
- * Session --> worker : context
+ * worker -> Manager : createContext()
+ * Manager --> worker : context
  * worker -> worker : context.access = context.kWrite
  * worker -> Manager : working_reference = preflight(["ref://asset"], context)[0]
  * Manager --> worker : "ref://asset=v4&state=inflight"
- * worker -> Session : freezeContext(context)
- * Session --> worker : token
+ * worker -> Manager : freezeContext(context)
+ * Manager --> worker : token
  * worker -> worker : write_state_file(token, {"ref://asset": working_entity_reference})
  * worker --> scheduler
- * deactivate Session
  * deactivate Manager
  * deactivate worker
  * ...
@@ -102,18 +98,16 @@
  * group distribute to n
  *     scheduler -> worker : someApp a_job.script ~--run ~--frame n --stateFile openassetio.json
  *     activate worker
- *     activate Session
  *     activate Manager
  *     worker -> worker : token, ref_map = read_state_file("openassetio.json")
- *     worker -> Session : thawContext(token)
- *     Session --> worker : context
+ *     worker -> Manager : thawContext(token)
+ *     Manager --> worker : context
  *     worker -> worker : context.access = context.kWrite
  *     worker -> worker : working_reference = ref_map["ref://asset"]
  *     worker -> Manager : [spec] = resolve([working_reference], {FileTrait.id}, context)
  *     Manager --> worker : Specification({"file": {"url": "file:///out.####.exr"}})
  *     worker -> worker : write_data(FileTrait(spec).url())
  *     worker --> scheduler
- *     deactivate Session
  *     deactivate Manager
  *     deactivate worker
  * end
@@ -121,11 +115,10 @@
  * == register task ==
  * scheduler -> worker : someApp a_job.script ~--register --stateFile openassetio.json
  * activate worker
- * activate Session
  * activate Manager
  * worker -> worker : token, ref_map = read_state_file("openassetio.json")
- * worker -> Session : thawContext(token)
- * Session --> worker : context
+ * worker -> Manager : thawContext(token)
+ * Manager --> worker : context
  * worker -> worker : context.access = context.kWrite
  * worker -> worker : working_reference = ref_map["ref://asset"]
  * worker -> Manager : path = resolve([working_reference], FileTrait.id, context)[0]
@@ -134,7 +127,6 @@
  * Manager --> worker : "ref://asset=v4"
  * worker -> worker : update_state_file({"ref://asset": final_reference})
  * worker --> scheduler
- * deactivate Session
  * deactivate Manager
  * deactivate worker
  * @enduml
