+++ tests/python/openassetio/hostAPI/test_session.py
@@ -25,7 +25,7 @@ from unittest import mock
 
 import pytest
 
-from openassetio import constants, exceptions, log, Context, TraitsData
+from openassetio import constants, exceptions, log
 from openassetio.hostAPI import Session, Manager, ManagerFactoryInterface
 
 
@@ -198,115 +198,6 @@ class Test_Session_currentManager:
         mock_manager_interface.mock.initialize.assert_called_once()
         mock_manager_interface.mock.setSettings.assert_called_once()
         assert mock_manager_interface.mock.setSettings.call_args[0][0] == some_settings
-
-
-class Test_Session_createContext:
-
-    def test_when_called_with_no_current_manager_then_a_RuntimeError_is_raised(
-            self, a_session, mock_manager_interface):
-
-        with pytest.raises(RuntimeError):
-            _ = a_session.createContext()
-
-        mock_manager_interface.mock.createState.assert_not_called()
-
-    def test_when_called_with_a_current_manager_then_context_is_created_with_expected_properties(
-            self, a_session, mock_manager_interface):
-
-        a_session.useManager("com.manager")
-
-        state_a = "state-a"
-        mock_manager_interface.mock.createState.return_value = state_a
-
-        context_a = a_session.createContext()
-
-        assert context_a.access == Context.kRead
-        assert context_a.retention == Context.kTransient
-        assert context_a.managerInterfaceState is state_a
-        assert context_a.locale is None
-        mock_manager_interface.mock.createState.assert_called_once()
-
-    def test_when_called_with_parent_then_props_copied_and_createState_called_with_parent_state(
-            self, a_session, mock_manager_interface):
-        # pylint: disable=protected-access
-
-        a_session.useManager("com.manager")
-
-        state_a = "state-a"
-        mock_manager_interface.mock.createState.return_value = state_a
-        context_a = a_session.createContext()
-        context_a.access = Context.kWrite
-        context_a.retention = Context.kSession
-        context_a.locale = TraitsData()
-        mock_manager_interface.mock.reset_mock()
-
-        state_b = "state-b"
-        mock_manager_interface.mock.createState.return_value = state_b
-        a_host_session = a_session.currentManager()._Manager__hostSession
-
-        context_b = a_session.createContext(parent=context_a)
-
-        assert context_b is not context_a
-        assert context_b.managerInterfaceState is state_b
-        assert context_b.access == context_a.access
-        assert context_b.retention == context_a.retention
-        assert context_b.locale == context_b.locale
-        mock_manager_interface.mock.createState.assert_called_once_with(
-            a_host_session, parentState=state_a)
-
-
-class Test_Session_freezeContext:
-
-    def test_when_called_with_no_manager_then_RuntimeError_is_raised(self, a_session):
-
-        with pytest.raises(RuntimeError):
-            a_session.freezeContext(Context())
-
-    def test_when_called_then_the_managers_frozen_token_is_returned(
-             self, a_session, mock_manager_interface):
-
-        expected_token = "a_frozen_token"
-        mock_manager_interface.mock.freezeState.return_value = expected_token
-
-        a_session.useManager("com.manager")
-
-        initial_state = "initial_state"
-        a_context = Context()
-        a_context.managerInterfaceState = initial_state
-
-        actual_token = a_session.freezeContext(a_context)
-
-        assert actual_token == expected_token
-
-        # pylint: disable=protected-access
-        a_host_session = a_session.currentManager()._Manager__hostSession
-        mock_manager_interface.mock.freezeState.assert_called_once_with(
-            initial_state, a_host_session)
-
-
-class Test_Session_thawContext:
-
-    def test_when_called_with_no_manager_then_RuntimeError_is_raised(self, a_session):
-
-        with pytest.raises(RuntimeError):
-            a_session.thawContext(Context())
-
-    def test_when_called_then_the_managers_thawed_state_is_set_in_the_context(
-             self, a_session, mock_manager_interface):
-
-        expected_state = "thawed_state"
-        mock_manager_interface.mock.thawState.return_value = expected_state
-
-        a_session.useManager("com.manager")
-
-        a_token = "frozen_token"
-        a_context = a_session.thawContext(a_token)
-
-        assert a_context.managerInterfaceState is expected_state
-
-        # pylint: disable=protected-access
-        a_host_session = a_session.currentManager()._Manager__hostSession
-        mock_manager_interface.mock.thawState.assert_called_once_with(a_token, a_host_session)
 
 
 class Test_Session_getSettings:
